Basic login structure in place
We have finished our login box, tied it to a database. We can now validate a username and password and allow a user into our member page  if correct.

We use sessions to check that the user is logged in before re-loading them. If they are not logged in, they are redirected to the normal home page.

Passwords are not encrypted yet and we have not established a log out button... But, we're getting there.