Password encrypted
The password is now encrypted and we have a start of a table in our member's area.

We still have to decide what we are doing with the normal nav functions when the member is logged in....

Then we can start to think about using a database to populate our columns.