We load some data!
We now use our news and events tables from within our database to populate the appropriate columns on our members page.

Before any data  is loaded, we check we have the user logged in with an active session.

Regardless of the order in the database, the items are listed in date order descending. This proved tricky because I converted the format to a UK date and this confused the ORDER BY so that we had to use STR_TO_DATE to fix that!

Things go up a level next because we want a pop-up with the article initiated  on a click on each article. The pop up isn't a major problem, getting it to start up PHP again and load the data may be tricky.. It's all parsed and PHP is done by the time we want this, we are asking something asynchronous from a synchronous process... maybe jQuery could help us? we'll have to see!