Meetings works
We now have meeting minutes in our member's area. It's basically the pdf files in a directory and the file name is used to help us order them by date using a user defined sort.

This name then is embedded into an anchor with PHP which when clicked, leads to the file.

At the moment, anyone can access this if they go to the directory and access the filename directly. There's options for this I might look at later!