Some security changes
We have restricted our directory listing within the mdata directory using htaccess. (on the live website I do this in more places).

I have a badlogin.html page to divert people to when they log in incorrectly. I could use this later with a session to track logins and block them if it became necessary to stop a brute force attack.